auto cache sections

+++ app/src/main/java/org/stepic/droid/model/Section.java
@@ -24,7 +24,7 @@ public class Section implements Serializable, Parcelable {
     private DateTimeFormatter mFormatForView;
 
     private int id;
-    private int course; // course id
+    private long course; // course id
     private long[] units;
     private int position;
     private String progress;
@@ -89,11 +89,11 @@ public class Section implements Serializable, Parcelable {
         this.id = id;
     }
 
-    public int getCourse() {
+    public long getCourse() {
         return course;
     }
 
-    public void setCourse(int course) {
+    public void setCourse(long course) {
         this.course = course;
     }
 
@@ -250,7 +250,7 @@ public class Section implements Serializable, Parcelable {
     @Override
     public void writeToParcel(Parcel dest, int flags) {
         dest.writeInt(this.id);
-        dest.writeInt(this.course);
+        dest.writeLong(this.course);
         dest.writeLongArray(this.units);
         dest.writeInt(this.position);
         dest.writeString(this.progress);
@@ -279,7 +279,7 @@ public class Section implements Serializable, Parcelable {
 
     private Section(Parcel in) {
         this.id = in.readInt();
-        this.course = in.readInt();
+        this.course = in.readLong();
         this.units = in.createLongArray();
         this.position = in.readInt();
         this.progress = in.readString();
